requirements.txt
b2btools
biopython
numpy
pandas
scipy
secstructartist